firewalldb: move bolt db logic to kvdb_store.go
In preparation for adding a DB struct in the db.go file which will be
db-backend agnostic.
